+++ src/main/java/com/yunche/loan/config/util/LockUtils.java
@@ -26,7 +26,7 @@ public class LockUtils {
      * @param key
      * @param timeOut 单位：微秒
      */
-    public boolean lock(String key, String val, Integer timeOut) {
+    public boolean lock(String key, String val, Long timeOut) {
 
         DefaultRedisScript<Long> redisScript = new DefaultRedisScript<>();
         redisScript.setScriptSource(new ResourceScriptSource(new ClassPathResource("lua/distributedLock.lua")));
